# static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "description": "Code With Prince",
      "version": "1.0.0",
      "title": "Code With Prince Flask API Tutorial",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "tags": [
      {
        "name": "Authentication API",
        "description": "Test API"
      }
    ],
     "servers": [
          {
          "url": "/"
          }
      ],
         "components": {
          "schemas": {
              "Register": {
                  "type": "object",
                  "properties": {
                      "client_id": {
                      "type": "string"
                      },
                      "client_secret": {
                      "type": "string"
                      },
                      "is_admin":{
                        "type":"boolean"
                      }
                  }
                },
                "Created":{
                "type" : "object",
                "properties" : {
                    "name" : {
                        "type" : "string"
                        }
                    }
                },
                "Authenticated":{
                "type": "object",
                "properties" :{
                    "credentials": {
                        "type": "string"
                        },
                    "expiration_seconde": {
                        "type": "integer"
                    },
                    "is_admin": {
                        "type": "boolean"
                    }
                }
                },
                "reponse":{
                    "type": "object",
                    "properties" :{
                        "client_id": {
                            "type": "string"
                            },
                        "client_secret": {
                            "type": "string"
                        }
                        
                    }
                },
              "login":{
                "type":"object",
                "properties":{
                    "client_id":{"type": "string"},
                    "client_secret":{"type": "string"}
                }
              }
          }
      },
       "paths": {
          "/register": {
              "post": {
                "tags": [
                    "Add a new User"
                ],
                "summary": "Add User",
                "requestBody": {
                "description": "Post New User To Database",
                "required": true,
                "content": {
                    "application/json": {
                    "schema": {
                        "$ref": "#/components/schemas/Register"
                        }
                    }
                }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                        "$ref": "#/components/schemas/Created"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                        "$ref": "#/components/schemas/Created"
                        }
                }
                
                },
                "parameters": [
                    {
                        "name": "client_id",
                        "in": "path",
                        "description": "todo id to update",
                        "required": true,
                        "schema":{
                            "type": "string"
                        } 
                    },
                        
                    {
                        "name": "client_secret",
                        "in": "path",
                        "description": "todo id to update",
                        "required": true,
                        "schema":{
                            "type": "string"
                        }   
                    },
                    {
                        "name": "is_admin",
                        "in": "path",
                        "description": "todo id to update",
                        "required": false,
                        "schema":{
                            "type": "boolean"
                        }
                        
                    }
                ]
              }
          },
          "/auth":{
  
            "post": {
                "tags": [
                    "Add a new User"
                ],
                "summary": "Add User",
                "requestBody": {
                "description": "Post New User To Database",
                "required": true,
                "content": {
                    "application/json": {
                    "schema": {
                        "$ref": "#/components/schemas/reponse"
                        }
                    }
                }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                        "$ref": "#/components/schemas/login"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                        "$ref": "#/components/schemas/login"
                        }
                }
                
                }
              }
          }
  }
  }